+++ Crud/src/main/java/com/KnowIt/Gym_intellect_Crud/Repository/Workout_planrepo.java
@@ -4,6 +4,8 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.KnowIt.Gym_intellect_Crud.Entity.WorkoutPlan;
 
-public interface Workout_planrepo extends JpaRepository<WorkoutPlan,Integer>{
+public interface Workout_planrepo extends JpaRepository<WorkoutPlan,Long>{
 	List<WorkoutPlan> findByPlanName(String planName);
+	
+//	List<WorkoutPlan> findByPlanName(String planName);
 }
